Added new structure

+++ libft/ft_abs.c
@@ -6,13 +6,13 @@
 /*   By: aszhilki <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/30 18:25:10 by aszhilki          #+#    #+#             */
-/*   Updated: 2019/11/30 18:26:23 by aszhilki         ###   ########.fr       */
+/*   Updated: 2020/01/31 15:06:05 by aszhilki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_abs(int number)
+float		ft_abs(float number)
 {
 	if (number >= 0)
 		return (number);
